pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "cat-cannon"
version = "0.1.0"
description = "Edge AI cat deterrence scaffold for Jetson Orin Nano"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.9.0",
  "pyright>=1.1.390",
]
bench = [
  "opencv-python>=4.10.0",
]
vision = [
  "ultralytics>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pyright]
include = ["src", "tests"]
typeCheckingMode = "basic"
